refactor: Refactor import statements to use the types folder instead of interfaces
